+++ app/src/main/java/com/sopt/now/compose/LoginActivity.kt
@@ -1,5 +1,6 @@
 package com.sopt.now.compose
 
+import LoginViewModel
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
@@ -21,6 +22,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 import androidx.compose.ui.res.stringResource
+import androidx.lifecycle.viewmodel.compose.viewModel
 import com.sopt.now.compose.TextField.CustomTextField
 import com.sopt.now.compose.api.ServicePool.authService
 import com.sopt.now.compose.dto.RequestLogInDto
@@ -39,7 +41,8 @@ class LoginActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    LoginScreen(intent)
+                    val loginViewModel: LoginViewModel = viewModel()
+                    LoginScreen(loginViewModel)
                 }
             }
         }
@@ -47,54 +50,8 @@ class LoginActivity : ComponentActivity() {
 }
 
 @Composable
-fun LoginScreen(intent: Intent) {
-    var login_id by remember { mutableStateOf("") }
-    var login_pw by remember { mutableStateOf("") }
+fun LoginScreen(loginViewModel: LoginViewModel) {
     val context = LocalContext.current
-
-    fun getLogInRequestDto(): RequestLogInDto {
-        val id = login_id
-        val password = login_pw
-        return RequestLogInDto(
-            authenticationId = id,
-            password = password
-        )
-    }
-
-    fun logIn() {
-        val loginRequest = getLogInRequestDto()
-        authService.logIn(loginRequest).enqueue(object : Callback<ResponseLogInDto> {
-            override fun onResponse(
-                call: Call<ResponseLogInDto>,
-                response: Response<ResponseLogInDto>,
-            ) {
-                if (response.isSuccessful) {
-                    val userId = response.headers()["location"]
-                    Toast.makeText(
-                        context,
-                        "로그인 성공",
-                        Toast.LENGTH_SHORT,
-                    ).show()
-
-                    val intent = Intent(context, MainActivity::class.java).apply {
-                        putExtra("userId", userId)
-                    }
-                    context.startActivity(intent)
-                } else {
-                    val error = response.message()
-                    Toast.makeText(
-                        context,
-                        "로그인 실패 $error",
-                        Toast.LENGTH_SHORT,
-                    ).show()
-                }
-            }
-
-            override fun onFailure(call: Call<ResponseLogInDto>, t: Throwable) {
-                Toast.makeText(context, "서버 에러 발생 ", Toast.LENGTH_SHORT).show()
-            }
-        })
-    }
 
     Column(
         modifier = Modifier
@@ -113,16 +70,16 @@ fun LoginScreen(intent: Intent) {
         Text(text = "ID", fontSize = 30.sp, color = Color.Black)
         Spacer(modifier = Modifier.height(16.dp))
         CustomTextField(
-            value = login_id,
-            onInputChange = { login_id = it },
+            value = loginViewModel.loginId.value,
+            onInputChange = { loginViewModel.loginId.value = it },
             label = stringResource(R.string.string_id_hint)
         )
 
         Spacer(modifier = Modifier.height(46.dp))
         Text(text = "PW", fontSize = 30.sp, color = Color.Black)
         CustomTextField(
-            value = login_pw,
-            onInputChange = { login_pw = it },
+            value = loginViewModel.loginPw.value,
+            onInputChange = { loginViewModel.loginPw.value = it },
             label = stringResource(R.string.string_pw_hint),
             isPwSecret = true
         )
@@ -146,7 +103,7 @@ fun LoginScreen(intent: Intent) {
                 }
                 Spacer(modifier = Modifier.width(10.dp))
                 Button(onClick = {
-                    logIn()
+                    loginViewModel.logIn(context)
                 }, modifier = Modifier.width(180.dp)) {
                     Text(text = stringResource(R.string.log_in_btn), fontSize = 20.sp)
                 }
